vinor/generator/templates/standard/pyproject.toml
[tool.poetry]
name = "standard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
standard = 'standard.apps.console.cli:run'


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
psycopg2-binary = "2.9.4"
sqlalchemy = "1.4.41"
pytest = "7.1.3"
pytest-schema = "0.1.1"
requests = "2.28.1"
flake8 = "5.0.4"
faker = "15.0.0"
python-multipart = "0.0.5"
pillow = "9.2.0"
pymysql = "1.0.2"
cryptography = "38.0.1"
jinja2 = "^3.1.2"
htmlmin = "^0.1.12"
typer = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
